refactor(web): split session-detail.tsx into focused components
session-detail.tsx: 1841 -> 492 lines (-73%). Extracted into
web/src/components/session-detail/:
- session-banners.tsx (477) - LinkRequest/Permission/Clipboard/AskQuestion
  banners + AskQuestionCard
- session-header.tsx (552) - collapsible info panel (status, model,
  tokens, cost, git, cache TTL, capabilities)
- session-tabs.tsx (225) - tab bar with permission-gated visibility
- session-input.tsx (225) - InputBar (memoized), DialogOverlay,
  ScrollToBottomButton

Pure structural move. Hooks order preserved, React.memo on InputBar
preserved, Zustand selectors untouched. Subagent detail panel + revive
button stayed inline (tight closure binding with parent state).
